A Theoretical-Empirical Approach to Estimating Sample Complexity of DNNs
Abstract
This paper focuses on understanding how the general-
ization error scales with the amount of the training data for
deep neural networks (DNNs). Existing techniques in sta-
tistical learning theory require a computation of capacity
measures, such as VC dimension, to provably bound this
error. It is however unclear how to extend these measures
to DNNs and therefore the existing analyses are applicable
to simple neural networks, which are not used in practice,
e.g., linear or shallow (at most two-layer) ones or other-
wise multi-layer perceptrons. Moreover many theoretical
error bounds are not empirically verifiable. In this paper
we derive estimates of the generalization error that hold
for deep networks and do not rely on unattainable capacity
measures. The enabling technique in our approach hinges
on two major assumptions: i) the network achieves zero
training error, ii) the probability of making an error on a
test point is proportional to the distance between this point
and its nearest training point in the feature space and at
certain maximal distance (that we call radius) it saturates.
Based on these assumptions we estimate the generalization
error of DNNs. The obtained estimate scales as O