Towards Interpretable and Transferable Speech Emotion Recognition: Latent Representation Based Analysis of Features, Methods and Corpora
Abstract
In recent years, speech emotion recognition (SER) has been
used in wide ranging applications, from healthcare to the com-
mercial sector. In addition to signal processing approaches,
methods for SER now also use deep learning techniques. How-
ever, generalizing over languages, corpora and recording condi-
tions is still an open challenge in the field. Furthermore, due to
the black-box nature of deep learning algorithms, a newer chal-
lenge is the lack of interpretation and transparency in the models
and the decision making process. This is critical when the SER
systems are deployed in applications that influence human lives.
In this work we address this gap by providing an in-depth anal-
ysis of the decision making process of the proposed SER sys-
tem. Towards that end, we present low-complexity SER based
on undercomplete- and denoising- autoencoders that achieve an
average classification accuracy of over 55% for four-class emo-
tion classification. Following this, we investigate the clustering
of emotions in the latent space to understand the influence of
the corpora on the model behavior and to obtain a physical in-
terpretation of the latent embedding. Lastly, we explore the role
of each input feature towards the performance of the SER.1
Index Terms: Speech emotion recognition, Autoencoder, In-
terpretability, Knowledge transfer, Transparency.