Addressing Annotation Imprecision for Tree Crown Delineation Using the RandCrowns Index
ABSTRACT:
Supervised methods for object delineation in remote sensing require labeled ground-truth data. Gathering sufficient high quality ground-
truth data is difficult, especially when the targets are of irregular shape or difficult to distinguish from the background or neighboring
objects.Tree crown delineation provides key information from remote sensing images for forestry, ecology, and management. However,
tree crowns in remote sensing imagery are often difficult to label and annotate due to irregular shape, overlapping canopies, shadowing,
and indistinct edges. There are also multiple approaches to annotation in this field (e.g., rectangular boxes vs. convex polygons) that
further contribute to annotation imprecision. However, current evaluation methods do not account for this uncertainty in annotations,
and quantitative metrics for evaluation can vary across multiple annotators. We address these limitations using an adaptation of the
Rand index for weakly-labeled crown delineation that we call RandCrowns. The RandCrowns metric reformulates the Rand index by
adjusting the areas over which each term of the index is computed to account for uncertain and imprecise object delineation labels.
Quantitative comparisons to the commonly used intersection over union (Jaccard similarity) method shows a decrease in the variance
generated by differences among multiple annotators. Combined with qualitative examples, our results suggest that this RandCrowns
metric is more robust for scoring target delineations in the presence of uncertainty and imprecision in annotations that are inherent
to tree crown delineation. Although the focus of this paper is on evaluation of tree crown delineations, annotation imprecision is a
challenge that is common across remote sensing of the environment (and many computer vision problems in general).