Stabilization of generative adversarial networks via noisy scale-space
Abstract
Generative adversarial networks (GAN) is a framework for generating fake data
based on given reals but is unstable in the optimization. In order to stabilize
GANs, the noise enlarges the overlap of the real and fake distributions at the cost
of significant variance. The data smoothing may reduce the dimensionality of
data but suppresses the capability of GANs to learn high-frequency information.
Based on these observations, we propose a data representation for GANs, called
noisy scale-space, that recursively applies the smoothing with noise to data in
order to preserve the data variance while replacing high-frequency information
by random data, leading to a coarse-to-fine training of GANs. We also present a
synthetic data-set using the Hadamard bases that enables us to visualize the true
distribution of data. We experiment with a DCGAN with the noise scale-space
(NSS-GAN) using major data-sets in which NSS-GAN overtook state-of-the-arts
in most cases independent of the image content.