Deep Extended Feedback Codes
Abstract—A new deep-neural-network (DNN) based error
correction encoder architecture for channels with feedback, called
Deep Extended Feedback (DEF), is presented in this paper.
The encoder in the DEF architecture transmits an information
message followed by a sequence of parity symbols which are
generated based on the message as well as the observations of the
past forward channel outputs sent to the transmitter through a
feedback channel. DEF codes generalize Deepcode [1] in several
ways: parity symbols are generated based on forward-channel
output observations over longer time intervals in order to provide
better error correction capability; and high-order modulation
formats are deployed in the encoder so as to achieve increased
spectral efficiency. Performance evaluations show that DEF codes
have better performance compared to other DNN-based codes for
channels with feedback.