EBIC.JL – an Efficient Implementation of Evolutionary Biclustering Algorithm in Julia
ABSTRACT
Biclustering is a data mining technique which searches for local
patterns in numeric tabular data with main application in bioin-
formatics. This technique has shown promise in multiple areas,
including development of biomarkers for cancer, disease subtype
identification, or gene-drug interactions among others. In this paper
we introduce EBIC.JL - an implementation of one of the most accu-
rate biclustering algorithms in Julia, a modern highly parallelizable
programming language for data science. We show that the new ver-
sion maintains comparable accuracy to its predecessor EBIC while
converging faster for the majority of the problems. We hope that
this open source software in a high-level programming language
will foster research in this promising field of bioinformatics and
expedite development of new biclustering methods for big data.