The Role of Time and Data: Online Conformance Checking in the Manufacturing Domain
Abstract
Process mining has matured as analysis instrument for process-oriented
data in recent years. Manufacturing is a challenging domain that craves
for process-oriented technologies to address digitalization challenges. We
found that process mining creates high expectations, but its implemen-
tation and usage by manufacturing experts such as process supervisors
and shopfloor workers remain unclear to a certain extent. Reason
1 is
that even though manufacturing allows for well-structured processes, the
actual workflow is rarely captured in a process model. Even if a model
is available, a software for orchestrating and logging the execution is of-
ten missing. Reason 2 refers to the work reality in manufacturing: a
process instance is started by a shopfloor worker who then turns to work
on other things. Hence continuous monitoring of the process instances
does not happen, i.e., process monitoring is merely a secondary task, and
the shopfloor worker can only react to problems/errors that have already
occurred.
1 and
2 motivate the goals of this study that is driven by
Technical Action Research (TAR). Based on the experimental artifact
TIDATE –a lightweight process execution and mining framework– it is
studied how the correct execution of process instances can be ensured
and how a data set suitable for process mining can be generated at run
time in a real-world setting. Secondly, it is investigated whether and how
process mining supports domain experts during process monitoring as a
secondary task. The findings emphasize the importance of online confor-
mance checking in manufacturing and show how appropriate data sets can
be identified and generated.