On the Pareto Front of Physics-Informed Neural Networks
ABSTRACT
Recently a new type of deep learning method has emerged, called physics-informed neural networks.
Despite their success in solving problems that are governed by partial differential equations, physics-
informed neural networks are often difficult to train. Frequently reported convergence issues are
still poorly understood and complicate the inference of correct system dynamics. In this paper, we
shed light on the training process of physics-informed neural networks. By trading between data-
and physics-based constraints in the network training, we study the Pareto front in multi-objective
optimization problems. We use the diffusion equation and Navier-Stokes equations in various test
environments to analyze the effects of system parameters on the shape of the Pareto front. Additionally,
we assess the effectiveness of state-of-the-art adaptive activation functions and adaptive loss weighting
methods. Our results demonstrate the prominent role of system parameters in the multi-objective
optimization and contribute to understanding convergence properties of physics-informed neural
networks.