Multi-modal Bifurcated Network for Depth Guided Image Relighting
Abstract
Image relighting aims to recalibrate the illumination set-
ting in an image. In this paper, we propose a deep learning-
based method called multi-modal bifurcated network (MB-
Net) for depth guided image relighting. That is, given an
image and the corresponding depth maps, a new image with
the given illuminant angle and color temperature is gener-
ated by our network. This model extracts the image and
the depth features by the bifurcated network in the encoder.
To use the two features effectively, we adopt the dynamic
dilated pyramid modules in the decoder. Moreover, to in-
crease the variety of training data, we propose a novel data
process pipeline to increase the number of the training data.
Experiments conducted on the VIDIT dataset show that the
proposed solution obtains the 1
st place in terms of SSIM
and PMS in the NTIRE 2021 Depth Guide One-to-one Re-
lighting Challenge.