Ethical Implementation of Artificial Intelligence to Select Embryos in In Vitro Fertilization
ABSTRACT
AI has the potential to revolutionize many areas of healthcare. 
Radiology, dermatology, and ophthalmology are some of the 
areas most likely to be impacted in the near future, and they 
have received significant attention from the broader research 
community. But AI techniques are now also starting to be used 
in in vitro fertilization (IVF), in particular for selecting which 
embryos to transfer to the woman. The contribution of AI to 
IVF is potentially significant, but must be done carefully and 
transparently, as the ethical issues are significant, in part 
because this field involves creating new people.
We first give a brief introduction to IVF and review the use of AI
for embryo selection. We discuss concerns with the 
interpretation of the reported results from scientific and 
practical perspectives. We then consider the broader ethical 
issues involved. We discuss in detail the problems that result 
from the use of black-box methods in this context and advocate 
strongly for the use of interpretable models. Importantly, there 
have been no published trials of clinical effectiveness, a problem 
in both the AI and IVF communities, and we therefore argue 
that clinical implementation at this point would be premature. 
Finally, we discuss ways for the broader AI community to 
become involved to ensure scientifically sound and ethically 
responsible development of AI in IVF.
CCS CONCEPTS
• Computing methodologies~Machine learning algorithms; 
Computer vision; Artificial intelligence; Machine learning; 
Machine learning approaches • Applied computing~Life and 
medical sciences.

KEYWORDS
IVF, In Vitro Fertilization, Embryo Selection, Artificial 
Intelligence, AI, Machine Learning, Interpretable, Black-Box,
Randomised Controlled Trials, RCT, Ethics