Unsupervised Anomaly Detection in MR Images using Multi-Contrast Information
Abstract
Anomaly detection in medical imaging is to distinguish the relevant biomarkers of diseases from those of normal 
tissues. Deep supervised learning methods have shown potentials in various detection tasks, but its performances 
would be limited in medical imaging fields where collecting annotated anomaly data is limited and labor-intensive. 
Therefore, unsupervised anomaly detection can be an effective tool for clinical practices, which uses only unlabeled 
normal images as training data. In this paper, we developed an unsupervised learning framework for pixel-wise 
anomaly detection in multi-contrast magnetic resonance imaging (MRI). The framework has two steps of feature 
generation and density estimation with Gaussian mixture model (GMM). A feature is derived through the learning 
of contrast-to-contrast translation that effectively captures the normal tissue characteristics in multi-contrast MRI. 
The feature is collaboratively used with another feature that is the low-dimensional representation of multi-contrast 
images. In density estimation using GMM, a simple but efficient way is introduced to handle the singularity problem 
which interrupts the joint learning process. The proposed method outperforms previous anomaly detection 
approaches. Quantitative and qualitative analyses demonstrate the effectiveness of the proposed method in anomaly 
detection for multi-contrast MRI