Searchable Hidden Intermediates for End-to-End Models of Decomposable Sequence Tasks
Abstract
End-to-end approaches for sequence tasks are
becoming increasingly popular. Yet for com-
plex sequence tasks, like speech translation,
systems that cascade several models trained
on sub-tasks have shown to be superior, sug-
gesting that the compositionality of cascaded
systems simplifies learning and enables so-
phisticated search capabilities. In this work,
we present an end-to-end framework that ex-
ploits compositionality to learn searchable hid-
den representations at intermediate stages of a
sequence model using decomposed sub-tasks.
These hidden intermediates can be improved
using beam search to enhance the overall per-
formance and can also incorporate external
models at intermediate stages of the network to
re-score or adapt towards out-of-domain data.
One instance of the proposed framework is
a Multi-Decoder model for speech translation
that extracts the searchable hidden intermedi-
ates from a speech recognition sub-task. The
model demonstrates the aforementioned bene-
fits and outperforms the previous state-of-the-
art by around +6 and +3 BLEU on the two test
sets of Fisher-CallHome and by around +3 and
+4 BLEU on the English-German and English-
French test sets of MuST-C.1