Learning data association without data association: An EM approach to neural assignment prediction
Abstract—Data association is a fundamental component of
effective multi-object tracking. Current approaches to data-
association tend to frame this as an assignment problem relying
on gating and distance-based cost matrices, or offset the challenge
of data association to a problem of tracking by detection. The
latter is typically formulated as a supervised learning problem,
and requires labelling information about tracked object identities
to train a model for object recognition. This paper introduces an
expectation maximisation approach to train neural models for
data association, which does not require labelling information.
Here, a Sinkhorn network is trained to predict assignment
matrices that maximise the marginal likelihood of trajectory
observations. Importantly, networks trained using the proposed
approach can be re-used in downstream tracking applications.