ICOS: Efficient and Highly Robust Point Cloud Registration with Correspondences
Point Cloud Registration is a fundamental problem in robotics and computer vision. Due to the limited accuracy in the matching process of 3D keypoints, the presence of outliers, probably in very large numbers, is common in many real-world applications. In this paper, we present ICOS (Inlier searching using COmpatible Structures), a novel, efficient and highly robust solution for the correspondence-based point cloud registration problem. Specifically, we (i) propose and construct a series of compatible structures for the registration problem where various invariants can be established, and (ii) design two time-efficient frameworks, one for known-scale registration and the other for unknown-scale registration, to filter out outliers and seek inliers from the invariant-constrained random sampling built upon the compatible structures. In this manner, even with extreme outlier ratios, inliers can be detected and collected for solving the optimal transformation, leading to our robust registration solver ICOS. Through plentiful experiments over standard real datasets, we demonstrate that: (i) our solver ICOS is fast, accurate, robust against as many as 99 whether the scale is known or unknown, outperforming other state-of-the-art methods, (ii) ICOS is practical for use in real-world applications.